Description: uint_8数据溢出修改 IssueNo:https://gitee.com/openharmony/arkui_ace_engine_lite/issues/I8L9EL Feature or Bugfix: Bugfix Binary Source: No Signed-off-by: huangji <huangji731@huawei.com>

+++ frameworks/src/core/modules/presets/cjson_parser.h
@@ -205,7 +205,7 @@ private:
     {
         uint64_t hash = 2166136261;
         uint32_t len = strlen(key);
-        for (uint8_t i = 0; i < len; i++) {
+        for (uint32_t i = 0; i < len; i++) {
             uint8_t keyVal = key[i];
             hash = (hash ^ keyVal) * NFV_PRIME;
         }
